Local notification is now scheduled when a new show is added and remove when it's deleted.
While the adding and removing notifications is working - I’ve yet to
test it. Also, we have to update the notification date every once in a
while (coz the notification date changes when the show starts to the
end date and vice versa). The end date is also sometimes estimated so
once I update the backend with the correct date - the notification for
that show has to be updated with the correct date as well.